test(connect): isolate DCode probe boundary coverage
